+++ CHANGELOG.md
@@ -33,6 +33,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `ActionsGenerator` will now generate C# warnings for non-private methods that are attributed as `YarnFunction` or `YarnCommand`.
 - `ActionsGenerator` still logs to a temporary location, but now into a `dev.yarnspinner.logs` folder inside the temporary location.
 - Auto-advancing `LineView`s will no longer attempt to advance dialogue that has been stopped.
+- `DialogueAdvanceInput` now supports Virtual Button names in addition to KeyCodes and Input Actions
+  - this can be configured to work on button or keycode release or press.
+  - defaults to on release.
 
 ### Removed